Add files via upload
Adds folders for the controllers, repositories, entities and services. 

Changes the `Review` class. Sets the `collection` attribute of the `@Document` annotation  to "reviews" instead of "movies". This way, and thanks to the @DocumentReference annotation of the `Movie` Class, the `createReview` method of the `ReviewService` class inserts a new review document in reviews collection and pushes an object containing the review’s id and body into the Movie document’s reviews array.